fix: 기존 API 명세가 달라짐에 따라 ResponseDTO 자료형 변경
- 유저의 스터디 성공률이 Int 에서 Double로 변경

+++ android/app/src/main/java/com/created/team201/data/mapper/StudyDetailMapper.kt
@@ -28,7 +28,7 @@ fun MemberResponseDto.toDomain(): Member = Member(
     id = this.id,
     tier = this.tier,
     nickname = this.nickname,
-    successRate = this.successRate,
+    successRate = this.successRate.toInt(),
     profileImage = this.profileImage ?: "",
 )
